Add close button in modal with backdrop
Closes #321
